+++ hutool-extra/pom.xml
@@ -327,6 +327,7 @@
 			<groupId>com.github.houbb</groupId>
 			<artifactId>pinyin</artifactId>
 			<version>0.2.0</version>
+			<optional>true</optional>
 		</dependency>
 		<dependency>
 			<groupId>cglib</groupId>
